--- src/main/java/BlackJack/Application.java
@@ -1,11 +1,0 @@
-package BlackJack;
-
-import BlackJack.controller.BlackjackController;
-
-public class Application {
-    public static void main(String[] args) {
-        BlackjackController blackjackController = new BlackjackController();
-        blackjackController.run();
-    }
-
-}
