--- src/main/java/BlackJack/domain/Card/Card.java
@@ -1,21 +1,0 @@
-package BlackJack.domain.Card;
-
-public class Card {
-
-    private Shape shape;
-    private Number number;
-
-    public Card(Shape shape, Number number) {
-        this.shape = shape;
-        this.number = number;
-    }
-
-    public Shape getShape() {
-        return shape;
-    }
-
-    public Number getNumber() {
-        return number;
-    }
-
-}
