--- src/main/java/BlackJack/domain/Card/Cards.java
@@ -1,49 +1,0 @@
-package BlackJack.domain.Card;
-
-import java.util.Collections;
-import java.util.List;
-
-public class Cards {
-    public static final int BUST_LINE = 21;
-    public static final int EXTRA_SCORE = 10;
-    private List<Card> deck;
-
-    public Cards(List<Card> deck) {
-        Collections.shuffle(deck);
-        this.deck = deck;
-    }
-
-    public List<Card> getDeck() {
-        return deck;
-    }
-
-    public void add(Card card) {
-        deck.add(card);
-    }
-
-    public int calculateScore() {
-        int score = deck.stream()
-                .mapToInt(card -> card.getNumber().getValue())
-                .sum();
-
-        return addAceScore(score);
-    }
-
-    private int addAceScore(int score) {
-        long countAce = deck.stream()
-                .filter(card -> card.getNumber().getDenomination().equals("A"))
-                .count();
-
-        for (int i = 0; i < countAce; i++) {
-            score = calculateAceScore(score);
-        }
-        return score;
-    }
-
-    private int calculateAceScore(int score) {
-        if (score + EXTRA_SCORE <= BUST_LINE) {
-            score += EXTRA_SCORE;
-        }
-        return score;
-    }
-}
