--- src/main/java/BlackJack/domain/User/Player.java
@@ -1,16 +1,0 @@
-package BlackJack.domain.User;
-
-import BlackJack.domain.Result;
-
-public class Player extends User {
-
-    private static final int PLAYER_ADD_CARD_LIMIT = 21;
-    public Player(String name) {
-        super(name);
-    }
-
-    public boolean isPossibleToAdd() {
-        return this.getScore() < PLAYER_ADD_CARD_LIMIT;
-    }
-
-}
