--- src/main/java/BlackJack/utils/ExeptionMessage.java
@@ -1,8 +1,0 @@
-package BlackJack.utils;
-
-public class ExeptionMessage {
-    public static final String INPUT_IS_NULL = "[ERROR] 입력은 null일 수 없습니다.";
-    public static final String INPUT_IS_BLANK = "[ERROR] 입력에 빈칸이 있을 수 없습니다.";
-    public static final String INPUT_IS_WRONG_FORMAT = "[ERROR] 입력은 y 혹은 n 이어야 합니다.";
-    public static final String CANNOT_ADD_CARD = "[ERROR] 카드를 더이상 받을 수 없습니다";
-}
