+++ src/main/java/blackJack/domain/Card/Card.java
@@ -1,0 +1,21 @@
+package blackJack.domain.Card;
+
+public class Card {
+
+    private Shape shape;
+    private Number number;
+
+    public Card(Shape shape, Number number) {
+        this.shape = shape;
+        this.number = number;
+    }
+
+    public Shape getShape() {
+        return shape;
+    }
+
+    public Number getNumber() {
+        return number;
+    }
+
+}
