+++ src/main/java/blackJack/domain/Card/Shape.java
@@ -1,0 +1,19 @@
+package blackJack.domain.Card;
+
+public enum Shape {
+    HEART("하트"),
+    DIAMOND("다이아몬드"),
+    SPADE("스페이드"),
+    CLOVER("클로버");
+
+    private String shapeName;
+
+    Shape(String shapeName) {
+        this.shapeName = shapeName;
+    }
+
+    public String getShapeName() {
+        return shapeName;
+    }
+
+}
