+++ src/main/java/blackJack/domain/User/Dealer.java
@@ -1,0 +1,26 @@
+package blackJack.domain.User;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class Dealer extends User {
+
+    private static final int DEALER_ADD_CARD_LIMIT = 16;
+
+    public Dealer() {
+        super("딜러");
+    }
+
+    public boolean checkBlackJack(){
+        List<String> openCardNumber = Arrays.asList("10","K","J","Q");
+        if(openCardNumber.contains(cards.getDeck().get(0).getNumber().getDenomination())){
+            return this.isBlackJack();
+        }
+        return false;
+    }
+
+
+    public boolean isPossibleToAdd() {
+        return this.getScore() < DEALER_ADD_CARD_LIMIT;
+    }
+}
