+++ src/main/java/blackJack/dto/PlayerResultDto.java
@@ -1,0 +1,28 @@
+package blackJack.dto;
+
+import blackJack.domain.Result;
+
+public class PlayerResultDto {
+
+    private String name;
+    private String result;
+
+    public PlayerResultDto(String name, String result) {
+        this.name = name;
+        this.result = result;
+    }
+
+    public static PlayerResultDto from(String name, Result result) {
+        return new PlayerResultDto(name, result.getValue());
+    }
+
+
+    public String getName() {
+        return name;
+    }
+
+    public String getResult() {
+        return result;
+    }
+}
+
